refactor : 패키지 구조를 수정한다
- input에 관련된 것을 console/input 밑으로 묶어서 응집도를 높여준다
