+++ src/common/config/AppConfig.java
@@ -4,8 +4,8 @@ import common.exception.GlobalExceptionHandler;
 import interfaces.console.controller.MainConsoleController;
 import interfaces.console.controller.ItineraryConsoleController;
 import interfaces.console.controller.TravelConsoleController;
-import interfaces.console.util.InputHandler;
-import interfaces.console.util.InputParser;
+import interfaces.console.input.InputHandler;
+import interfaces.console.input.InputParser;
 import interfaces.console.util.RetryHandler;
 import interfaces.console.view.TravelView;
 import interfaces.console.view.ItineraryView;
