+++ src/common/exception/ErrorCode.java
@@ -35,10 +35,6 @@ public enum ErrorCode {
         this.message = message;
     }
 
-    public String getCode() {
-        return code;
-    }
-
     public String getMessage() {
         return message;
     }
